// src/types/Location.ts
export type Location = {
    id: number;
    name: string;
    street: string;
    city: string;
    state: string;
    zip: string;
    phone: string;
    lat: string;
    lon: string;
    website: string;
    created_at: string;
    updated_at: string;
    zone_id: number | null;
    region_id: number;
    location_type_id: number;
    description: string;
    operator_id: number | null;
    date_last_updated: string;
    last_updated_by_user_id: number;
    is_stern_army: boolean | null;
    country: string;
    ic_active: boolean | null;
    distance: number;
    bearing: string;
    machine_names: string[];
    machine_ids: number[];
    num_machines: number;
};